Renamed appointment.controller.test.js to appointmentTest.js for consistency
